Fix inheriting from WebSocket when using multiple inheritance
If super(...).__init__ is not called in WebSocket.__init__, then the chain of execution ends here when calling super(...).__init__ from a child class which inherits WebSocket. This may be fine when using single inheritance, but will prevent __init__ of other classes inherited by the child class from being called.